pyproject.toml
[project]
name = "qabrains-ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest>=2.15.3",
    "faker>=40.1.2",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
]
